Backend/requirements.txt
django
djangorestframework
psycopg2-binary
ultralytics==8.3.0
opencv-python
torch
numpy
scikit-learn
scipy